requirements.txt
anyio==3.6.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.79.0
google-api-core==2.8.2
google-auth==2.11.0
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
h11==0.13.0
idna==3.3
Pillow==9.2.0
protobuf==4.21.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.2
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tensorflow==2.9.1
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.2
